backend/requirements.txt
agno
groq
pypdf
fastapi
uvicorn[standard]
python-dotenv
sentence-transformers
pinecone==5.3.1
pypdf2
